change of product and order table.

+++ src/main/java/koala/sloth/demo/respository/OrderRepository.java
@@ -3,6 +3,6 @@ package koala.sloth.demo.respository;
 import koala.sloth.demo.domain.Order;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface OrderRepository extends JpaRepository<Order,Long> {
+public interface OrderRepository extends JpaRepository<Order,Integer> {
 
 }
